plugins/cisco_firepower_management_center/requirements.txt
# List third-party dependencies here, separated by newlines. 
# All dependencies must be version-pinned, eg. requests==1.2.0
# See: https://pip.pypa.io/en/stable/user_guide/#requirements-files
fmcapi==20211214.0
validators==0.18.2
pyOpenSSL==21.0.0
parameterized==0.8.1